requirements.txt
bleach==3.3.1
bokeh==2.3.3
pandas
casatasks==6.2.0.124
casatestutils==6.2.0.124
casatools==6.2.0.124
casaviewer==1.2.1a13492.dev1
casadata
certifi==2021.5.30
charset-normalizer==2.0.3
cycler==0.10.0
grpcio==1.39.0
idna==3.2
importlib-metadata==4.6.1
Jinja2==3.0.1
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
numpy==1.19.5
packaging==21.0
panel==0.12.0
param==1.11.1
Pillow==8.3.1
protobuf==3.17.3
pyct==0.4.8
pyparsing==2.4.7
python-dateutil==2.8.2
pyviz-comms==2.1.0
PyYAML==5.4.1
requests==2.26.0
scipy==1.5.4
six==1.16.0
tornado==6.1
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
webencodings==0.5.1
zipp==3.5.0
